lidc_unet/requirements.txt
# LIDC-IDRI U-Net 肺结节分割项目依赖 (PyTorch 版本)

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 图像处理
Pillow>=9.0.0
scikit-image>=0.19.0
opencv-python>=4.5.0

# 数据处理
numpy>=1.21.0
pandas>=1.4.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 机器学习工具
scikit-learn>=1.0.0

# 进度条
tqdm>=4.62.0

# 可选：模型摘要（类似 Keras model.summary()）
# torchinfo>=1.7.0

# 可选：TensorBoard 可视化
# tensorboard>=2.10.0
